refactor: fix some merge conflict with variable names
