+++ src/main/java/com/distinction/barbenheimer/service/OngoingPurchaseService.java
@@ -1,0 +1,7 @@
+package com.distinction.barbenheimer.service;
+
+public interface OngoingPurchaseService {
+
+    public String createCustomerIdentifyingToken();
+
+}
